debug: add detailed decode logging + fix format string warnings
- APIService: loggedDecode helper prints exact failing key/path and
  raw JSON to console on DecodingError (look for 🔴 [API] in Xcode log)
- Fix %.0f%% format string warnings with explicit Double casts in
  ProdukDetailView and BerandaView

+++ oura studio frontend/Screens/Beranda/BerandaView.swift
@@ -129,7 +129,7 @@ struct BerandaView: View {
                     Spacer()
                     statPill(
                         icon: "chart.line.uptrend.xyaxis",
-                        label: String(format: "%.0f%% margin", dashboard?.avgMarginPct ?? 0)
+                        label: String(format: "%.0f%% margin", (dashboard?.avgMarginPct ?? 0) as Double)
                     )
                 }
             }
